LayoutTests/http/tests/websocket/tests/client-close.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../../../js-test-resources/js-test-style.css">
<script src="../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("WebSocket: Test client-initiated close.");

window.jsTestIsAsync = true;

var ws = new WebSocket("ws://127.0.0.1:8880/websocket/tests/client-close");
var closeEvent;

ws.onopen = function()
{
    debug("Connected");
    ws.close();
};

ws.onmessage = function(messageEvent)
{
    debug("Received: " + messageEvent.data);
};

ws.onclose = function(event)
{
    debug("Closed");
    closeEvent = event;
    shouldBeTrue("closeEvent.wasClean");
    finishJSTest();
};

var successfullyParsed = true;
</script>
<script src="../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
